#63 - refactor: 메시지 Service 단일 책임 원칙 적용
