+++ src/main/java/com/gxdxx/instagram/service/message/MessageCreateService.java
@@ -1,16 +1,11 @@
-package com.gxdxx.instagram.service;
+package com.gxdxx.instagram.service.message;
 
-import com.gxdxx.instagram.dto.request.MessageListRequest;
 import com.gxdxx.instagram.dto.request.MessageSendRequest;
-import com.gxdxx.instagram.dto.response.MessageListResponse;
-import com.gxdxx.instagram.dto.response.MessageResponse;
 import com.gxdxx.instagram.dto.response.SuccessResponse;
 import com.gxdxx.instagram.entity.ChatRoom;
 import com.gxdxx.instagram.entity.Message;
 import com.gxdxx.instagram.entity.User;
-import com.gxdxx.instagram.exception.ChatRoomNotFoundException;
 import com.gxdxx.instagram.exception.InvalidRequestException;
-import com.gxdxx.instagram.exception.UnauthorizedAccessException;
 import com.gxdxx.instagram.exception.UserNotFoundException;
 import com.gxdxx.instagram.repository.ChatRoomRepository;
 import com.gxdxx.instagram.repository.MessageRepository;
@@ -20,17 +15,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Map;
 
-@RequiredArgsConstructor
 @Transactional
+@RequiredArgsConstructor
 @Service
-public class MessageService {
+public class MessageCreateService {
 
     private final MessageRepository messageRepository;
-    private final UserRepository userRepository;
     private final ChatRoomRepository chatRoomRepository;
+    private final UserRepository userRepository;
 
     public SuccessResponse sendMessage(MessageSendRequest request, String nickname) {
         User sendUser = getUserByNickname(nickname);
@@ -63,23 +56,6 @@ public class MessageService {
     private ChatRoom getOrCreateChatRoom(User sendUser, User receiveUser) {
         return chatRoomRepository.findByUserIds(sendUser.getId(), receiveUser.getId())
                 .orElseGet(() -> chatRoomRepository.save(ChatRoom.of(sendUser, receiveUser)));
-    }
-
-    @Transactional(readOnly = true)
-    public MessageResponse getMessages(MessageListRequest request, String nickname) {
-        User requestUser = getUserByNickname(nickname);
-        ChatRoom chatRoom = chatRoomRepository.findById(request.chatRoomId())
-                .orElseThrow(ChatRoomNotFoundException::new);
-        if (!chatRoom.hasUser(requestUser)) {
-            throw new UnauthorizedAccessException();
-        }
-        Long cursor = (request.cursor() == null)
-                ? messageRepository.findMaxMessageId().map(maxId -> maxId + 1).orElse(0L)
-                : request.cursor();
-        List<MessageListResponse> messages = messageRepository.getMessagesByCursor(requestUser.getId(), chatRoom.getId(), cursor, 5);
-        Long nextCursor = messages.isEmpty() ? 0L : messages.get(messages.size() - 1).getMessageId();
-
-        return MessageResponse.of(nextCursor, messages);
     }
 
 }
